NO JIRA. Ensure SHA-1 manifest is present before it is used.
