+++ src/test/java/nl/knaw/dans/ingest/core/service/ManifestHelperImplTest.java
